Dev -> Master

+++ todo-backend-springboot/src/main/kotlin/epo/todo/backend/service/TodoElementService.kt
@@ -2,8 +2,10 @@ package epo.todo.backend.service
 
 import epo.todo.backend.domain.TodoElementEntity
 import epo.todo.backend.exception.BadRequestException
+import epo.todo.backend.exception.NotFoundException
 import epo.todo.backend.model.TodoElementDto
 import epo.todo.backend.repository.TodoElementRepository
+import mu.KotlinLogging
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -11,12 +13,16 @@ import org.springframework.transaction.annotation.Transactional
 @Transactional
 class TodoElementService(val todoElementRepository: TodoElementRepository) {
 
+    private val log = KotlinLogging.logger {}
+
     fun create(todoElementDto: TodoElementDto): TodoElementDto {
-        if (todoElementRepository.findByCategoryAndText(todoElementDto.category,
-                                                        todoElementDto.text).isNotEmpty()) {
-            throw BadRequestException("A todo element with the same category and text already exists.")
+        if (todoElementRepository.existsByCategoryAndText(todoElementDto.category,
+                                                          todoElementDto.text)) {
+            throw BadRequestException("A todo element with category=${todoElementDto.category} " +
+                                              "and text=${todoElementDto.text} already exists.")
         }
         return upsert(todoElementDto)
+                .also { log.info { "Created todo element=$it" } }
     }
 
     fun update(id: Int, todoElementDto: TodoElementDto): TodoElementDto {
@@ -28,15 +34,18 @@ class TodoElementService(val todoElementRepository: TodoElementRepository) {
         }
 
         return upsert(todoElementDto)
+                .also { log.info { "Updated todo element=$todoElementDto" } }
     }
 
     fun get(category: String?): List<TodoElementDto> {
+        log.info { "Getting todos with category=$category" }
         val todoElements: List<TodoElementEntity> =
                 when {
                     category == null -> todoElementRepository.findAll()
                     else -> todoElementRepository.findAllByCategory(category)
                 }
         return todoElements.map { TodoElementDto(it) }
+                .also { log.info { "get by category=$category returned todos=$it" } }
     }
 
     private fun upsert(todoElementDto: TodoElementDto): TodoElementDto {
@@ -45,6 +54,11 @@ class TodoElementService(val todoElementRepository: TodoElementRepository) {
     }
 
     fun delete(id: Int) {
+        if (!todoElementRepository.existsById(id)) {
+            throw NotFoundException("Could not delete todo with id=$id as it does not exist.")
+        }
+
         todoElementRepository.deleteById(id)
+                .also { log.info { "Deleted todo with id=$id" } }
     }
 }
